requirements.txt
openai>=1.12.0
google-generativeai>=0.3.0
anthropic>=0.19.0
llama-cpp-python>=0.2.6
python-dotenv>=1.0.0
aiohttp>=3.9.0